fix: correct import paths in moved test files, restore auth-security compat
- Changed require('../lib/commands/...') to require('../../lib/commands/...')
  in 6 test files moved from test/ to test/commands/
- Added 'node scripts/' as acceptable CLI pattern in auth-security test
  so lefthook.yml's check-forge-token.js references pass OWASP A07 check
